pyproject.toml
[tool.poetry]
name        = "news-scraper"
version     = "0.1.0"
description = "A scheduler to get news from yfinance."

packages = [
  { include = "application", from = "src" },
  { include = "domain", from = "src" },
  { include = "infrastructure", from = "src" },
]

[tool.poetry.dependencies]
python                    = ">=3.9,<3.13"

# API server
fastapi                   = "^0.95.0"
uvicorn                   = { extras = ["standard"], version = "^0.18.0" }

# Data fetchers & HTTP
yfinance                  = "^0.2.18"
requests                  = "^2.26.0"
httpx                     = "^0.24.0"
anyio                     = ">=3.4.0,<5"

# Config & validation
pydantic                  = "^1.10.7"
toml                      = "^0.10.2"

# Persistence & messaging
pymongo                   = "^4.6.1"
google-cloud-pubsub       = "^2.19.1"

# Model / ML
numpy                     = ">=1.24,<2.0"
pandas                    = "^2.0.3"

apscheduler  = "^3.11"

[tool.poetry.group.dev.dependencies]
pytest                    = "^7.0.0"
black                     = "^24.3.0"
mypy                      = "^1.5.1"
optuna                    = "^3.6.0"

[tool.black]
line-length    = 88
target-version = ["py311"]
include        = "\\.pyi?$"
exclude        = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[build-system]
requires      = ["poetry-core>=1.0.0", "setuptools>=42", "wheel"]
build-backend = "poetry.core.masonry.api"
